ourbit_bot/requirements.txt
anyio==4.4.0
asyncio==3.4.3
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.53.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pymongo==4.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.4
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tzdata==2024.1
urllib3==2.2.2
websockets==12.0
Werkzeug==3.0.3
